move method models + notifications page
