+++ mapito/TableCellNotif.swift
@@ -1,0 +1,14 @@
+//
+//  TableCellNotif.swift
+//  mapito
+//
+//  Created by m2sar on 01/06/2018.
+//  Copyright © 2018 code2geny. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class TableCellNotif: UITableViewCell {
+    
+}
